requirements.txt
tensorflow==2.12.0
numpy
pillow
streamlit
opencv-python
matplotlib
seaborn
scikit-learn

